Fix plugin.spec.ts.
