+++ packages/server/src/api/controllers/plugin/utils.ts
@@ -6,12 +6,7 @@ export async function downloadUnzipTarball(
   name: string,
   headers = {}
 ) {
-  try {
-    const path = createTempFolder(name)
-    await objectStore.downloadTarballDirect(url, path, headers)
-
-    return path
-  } catch (e: any) {
-    throw new Error(e.message)
-  }
+  const path = createTempFolder(name)
+  await objectStore.downloadTarballDirect(url, path, headers)
+  return path
 }
