+++ packages/server/src/utilities/fileSystem/filesystem.ts
@@ -3,7 +3,7 @@ import fsp from "fs/promises"
 import { budibaseTempDir } from "../budibaseDir"
 import { join } from "path"
 import env from "../../environment"
-import tarStream from "tar-stream"
+import tar from "tar-stream"
 import zlib from "zlib"
 import { pipeline } from "stream/promises"
 
@@ -155,46 +155,29 @@ export const createTempFolder = (item: any) => {
   return path
 }
 
-export const extractTarball = async (fromFilePath: string, toPath: string) => {
-  const extract = tarStream.extract()
+export const extractTarball = async (from: string, to: string) => {
+  const extract = tar.extract()
 
-  extract.on("entry", async (header, stream, next) => {
-    const targetPath = join(toPath, header.name)
+  let read: Readable = fs.createReadStream(from)
+  if (await isGzip(read)) {
+    read = read.pipe(zlib.createGunzip())
+  }
+  read.pipe(extract)
 
-    if (header.type === "directory") {
-      try {
-        await fs.promises.mkdir(targetPath, { recursive: true })
-      } catch (err: any) {
-        if (err.code !== "EEXIST") throw err
-      }
-      stream.resume()
-      next()
-    } else if (header.type === "file") {
-      // Ensure directory exists
-      const dir = join(targetPath, "..")
-      try {
-        await fs.promises.mkdir(dir, { recursive: true })
-      } catch (err: any) {
-        if (err.code !== "EEXIST") throw err
-      }
+  for await (const entry of extract) {
+    const {
+      header: { name, type },
+    } = entry
+    const path = join(to, name)
 
-      const writeStream = fs.createWriteStream(targetPath)
-      stream.pipe(writeStream)
-      stream.on("end", next)
-    } else {
-      stream.resume()
-      next()
+    if (type === "directory") {
+      await fs.promises.mkdir(path, { recursive: true })
+    } else if (type === "file" || type === "contiguous-file") {
+      await fs.promises.mkdir(join(path, ".."), { recursive: true })
+      await fsp.writeFile(path, entry, "binary")
     }
-  })
 
-  const readStream = fs.createReadStream(fromFilePath)
-  const isGzipped =
-    fromFilePath.endsWith(".gz") || fromFilePath.endsWith(".tgz")
-
-  if (isGzipped) {
-    await pipeline(readStream, zlib.createGunzip(), extract)
-  } else {
-    await pipeline(readStream, extract)
+    entry.resume()
   }
 }
 
@@ -229,4 +212,92 @@ export const deleteFolderFileSystem = (path: PathLike) => {
   }
 
   fs.rmSync(path, { recursive: true, force: true })
+}
+
+/**
+ * Peek at the next chunk of data in the stream without consuming it.
+ * @param input The input stream to read from.
+ * @param length The number of bytes to peek.
+ * @returns A promise that resolves to the peeked data.
+ */
+export async function peek(input: Readable, length: number): Promise<Buffer> {
+  // Try synchronous read first
+  const immediateBuffer = input.read(length)
+  if (immediateBuffer) {
+    input.unshift(immediateBuffer)
+    return immediateBuffer
+  }
+
+  // If not available immediately, we need to collect chunks
+  return new Promise((resolve, reject) => {
+    const chunks: Uint8Array[] = []
+    let totalLength = 0
+
+    const cleanup = () => {
+      input.removeListener("readable", onReadable)
+      input.removeListener("end", onEnd)
+      input.removeListener("error", onError)
+    }
+
+    const onReadable = () => {
+      let chunk: Buffer
+
+      while ((chunk = input.read()) !== null) {
+        chunks.push(new Uint8Array(chunk))
+        totalLength += chunk.length
+
+        // If we have enough data, stop reading
+        if (totalLength >= length) {
+          cleanup()
+
+          // Put all chunks back in reverse order
+          for (let i = chunks.length - 1; i >= 0; i--) {
+            input.unshift(Buffer.from(chunks[i]))
+          }
+
+          const combined = Buffer.concat(chunks)
+          resolve(combined.subarray(0, length))
+          return
+        }
+      }
+
+      // If we get here, we don't have enough data yet
+      // The readable event will fire again when more data is available
+    }
+
+    const onEnd = () => {
+      cleanup()
+
+      if (chunks.length === 0) {
+        reject(new Error("Stream ended with no data"))
+        return
+      }
+
+      // Stream ended, can't put data back, just return what we could read
+      const combined = Buffer.concat(chunks)
+      resolve(combined)
+    }
+
+    const onError = (err: Error) => {
+      cleanup()
+      reject(err)
+    }
+
+    input.on("readable", onReadable)
+    input.on("end", onEnd)
+    input.on("error", onError)
+
+    // Try reading immediately in case data is already buffered
+    onReadable()
+  })
+}
+
+export async function isGzip(input: Readable) {
+  return (await peek(input, 2)).toString("hex") === "1f8b" // Gzip magic number
+}
+
+export async function isTar(input: Readable) {
+  const chunk = await peek(input, 262)
+  const magic = chunk.toString("utf8", 257, 262)
+  return magic === "ustar"
 }
